Foundation.Check

Description

An implementation of a test framework and property expression & testing

Documentation

data Gen a

Generator monad

class Arbitrary a where

How to generate an arbitrary value for a

Methods

arbitrary :: Gen a

oneof :: NonEmpty [Gen a] -> Gen a

elements :: NonEmpty [a] -> Gen a

frequency :: NonEmpty [(Word, Gen a)] -> Gen a

Call one of the generator weighted

between :: (Word, Word) -> Gen Word

data Test where

different type of tests supported

Constructors

Unit :: String -> IO () -> Test

Property :: IsProperty prop => String -> prop -> Test

Group :: String -> [Test] -> Test

CheckPlan :: String -> Check () -> Test

testName :: Test -> String

Name of a test

Property

data PropertyCheck

The type of check this test did for a property

data Property

Constructors

Prop

Fields

unProp :: Gen PropertyTestArg

class IsProperty p where

Methods

property :: p -> Property

(===) :: (Show a, Eq a, Typeable a) => a -> a -> PropertyCheck infix 4

A property that check for equality of its 2 members.

propertyCompare

Arguments

:: (Show a, Typeable a)

=> String
name of the function used for comparaison, e.g. (<)

-> (a -> a -> Bool)
function used for value comparaison

-> a
value left of the operator

-> a
value right of the operator

-> PropertyCheck

A property that check for a specific comparaison of its 2 members.

This is equivalent to === but with compare

propertyCompareWith

Arguments

:: String
name of the function used for comparaison, e.g. (<)

-> (a -> a -> Bool)
function used for value comparaison

-> (a -> String)
function used to pretty print the values

-> a
value left of the operator

-> a
value right of the operator

-> PropertyCheck

A property that check for a specific comparaison of its 2 members.

This is equivalent to === but with compare and a given method to pretty print the values.

propertyAnd :: PropertyCheck -> PropertyCheck -> PropertyCheck

A conjuctive property composed of 2 properties that need to pass

propertyFail :: String -> PropertyCheck

forAll :: (Show a, IsProperty prop) => Gen a -> (a -> prop) -> Property

Running a generator for a specific type under a property

Check Plan

data Check a

validate :: IsProperty prop => String -> prop -> Check ()

pick :: String -> IO a -> Check a

iterateProperty :: CountOf TestResult -> GenParams -> (Word64 -> GenRng) -> Property -> IO (PropertyResult, CountOf TestResult)