Foundation.Class.Bifunctor

Description

Formally, the class Bifunctor represents a bifunctor from Hask -> Hask.

Intuitively it is a bifunctor where both the first and second arguments are covariant.

You can define a Bifunctor by either defining bimap or by defining both first and second.

Documentation

module Basement.Compat.Bifunctor