Foundation.Exception

Documentation

finally :: MonadBracket m => m a -> m b -> m a

try :: (MonadCatch m, Exception e) => m a -> m (Either e a)

data SomeException

The SomeException type is the root of the exception type hierarchy. When an exception of type e is thrown, behind the scenes it is encapsulated in a SomeException.