Foundation.Format.CSV

Description

Provies the support for Comma Separated Value

CSV

data CSV

CSV Type

Builder

String Bulider

csvStringBuilder :: CSV -> Builder

serialise the CSV document into a UTF8 string

rowStringBuilder :: Row -> Builder

fieldStringBuilder :: Field -> Builder

Block Builder

csvBlockBuilder :: CSV -> Builder

serialise the CSV document into a UTF8 encoded (Block Word8)

rowBlockBuilder :: Row -> Builder

fieldBlockBuilder :: Field -> Builder

Conduit

rowC :: (Record row, Monad m) => Conduit row (Block Word8) m ()

Parser

String Bulider

file :: Parser String CSV

record :: Parser String Row

record_ :: forall row. (Typeable row, Record row) => Parser String row

field :: Parser String Field

Conduit

recordC :: (Monad m, MonadThrow m) => Conduit String Row m ()

Row

data Row

CSV Row

class Record a where

Methods

toRow :: a -> Row

fromRow :: Row -> Either String a

Field

data Field

CSV field

Constructors

FieldInteger Integer

FieldDouble Double

FieldString String Escaping

data Escaping

Constructors

NoEscape

Escape

DoubleEscape

class IsField a where

Methods

toField :: a -> Field

fromField :: Field -> Either String a

helpers

integral :: Into Integer a => a -> Field

helper function to create a FieldInteger

float :: Double -> Field

string :: String -> Field

heler function to create a FieldString.

This function will findout automatically if an escaping is needed. if you wish to perform the escaping manually, do not used this function