Foundation.Monad

Documentation

class Monad m => MonadIO (m :: Type -> Type) where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad. This allows us to run IO computations in any monadic stack, so long as it supports these kinds of operations (i.e. IO is the base monad for the stack).

Example

Expand

import Control.Monad.Trans.State -- from the "transformers" library

printState :: Show s => StateT s IO ()
printState = do
  state <- get
  liftIO $ print state

Had we omitted liftIO, we would have ended up with this error:

• Couldn't match type ‘IO’ with ‘StateT s IO’
 Expected type: StateT s IO ()
   Actual type: IO ()

The important part here is the mismatch between StateT s IO () and IO ().

Luckily, we know of a function that takes an IO a and returns an (m a): liftIO, enabling us to run the program and see the expected results:

> evalStateT printState "hello"
"hello"

> evalStateT printState 3
3

class Monad m => MonadFailure (m :: Type -> Type) where

Monad that can represent failure

Similar to MonadFail but with a parametrized Failure linked to the Monad

Associated Types

type Failure (m :: Type -> Type)

The associated type with the MonadFailure, representing what failure can be encoded in this monad

Methods

mFail :: Failure m -> m ()

Raise a Failure through a monad.

class Monad m => MonadThrow m where

Monad that can throw exception

Methods

throw :: Exception e => e -> m a

Throw immediatity an exception. Only a MonadCatch monad will be able to catch the exception using catch

class MonadThrow m => MonadCatch m where

Monad that can catch exception

Methods

catch :: Exception e => m a -> (e -> m a) -> m a

class MonadCatch m => MonadBracket m where

Monad that can ensure cleanup actions are performed even in the case of exceptions, both synchronous and asynchronous. This usually excludes continuation-based monads.

Methods

generalBracket

Arguments

:: m a
acquire some resource

-> (a -> b -> m ignored1)
cleanup, no exception thrown

-> (a -> SomeException -> m ignored2)
cleanup, some exception thrown. The exception will be rethrown

-> (a -> m b)
inner action to perform with the resource

-> m b

A generalized version of the standard bracket function which allows distinguishing different exit cases.

class MonadTrans trans where

Basic Transformer class

Methods

lift :: Monad m => m a -> trans m a

Lift a computation from an inner monad to the current transformer monad

newtype Identity a

Identity functor and monad. (a non-strict monad)

Since: base-4.8.0.0

Constructors

Identity

Fields

runIdentity :: a

replicateM :: Applicative m => CountOf a -> m a -> m [a]

replicateM n act performs the action n times, gathering the results.