Foundation.Network.HostName

Description

Documentation

newtype HostName

HostName

Constructors

HostName

Fields

toString :: String

data HostNameInfo address_type

HostName Info

Constructors

HostNameInfo

Fields

officialName :: !HostName

official names

aliases :: !(Array HostName)

known aliases

addresses :: !(Array address_type)

known addresses

getHostNameInfo :: (Eq address_type, Storable address_type, SocketFamily address_type) => HostName -> IO (HostNameInfo address_type)

get HostName info:

retrieve the official name, the aliases and the addresses associated to this hostname.

For cross-platform compatibility purpose, this function is using a *C* non re-entrant function gethostbyname2. This function is using a `MVar ()` to avoid a race condition and should be safe to use.

getHostNameInfo_ :: (SocketFamily address_type, Eq address_type, Storable address_type) => Proxy address_type -> HostName -> IO (HostNameInfo address_type)

like getHostNameInfo but takes a Proxy to help with the type checker.