Foundation.Parser

Description

The current implementation is mainly, if not copy/pasted, inspired from memory's Parser.

Foundation Parser makes use of the Foundation's Collection and Sequential classes to allow you to define generic parsers over any Sequential of inpu.

This way you can easily implements parsers over LString, String.

flip parseOnly "my.email@address.com" $ do
   EmailAddress
     <$> (takeWhile ((/=) '@' <*  element '@')
     <*> takeAll

Documentation

data Parser input result

Foundation's Parser monad.

Its implementation is based on the parser in memory.

parse :: ParserSource input => Parser input a -> input -> Result input a

Run a Parser on a ByteString and return a Result

parseFeed :: (ParserSource input, Monad m) => m (Chunk input) -> Parser input a -> input -> m (Result input a)

Run a parser on an @initial input.

If the Parser need more data than available, the @feeder function is automatically called and fed to the More continuation.

parseOnly :: (ParserSource input, Monoid (Chunk input)) => Parser input a -> input -> Either (ParseError input) a

parse only the given input

The left-over `Element input` will be ignored, if the parser call for more data it will be continuously fed with Nothing (up to 256 iterations).

Result

data Result input result

result of executing the parser over the given input

Constructors

ParseFailed (ParseError input)
the parser failed with the given ParserError

ParseOk (Chunk input) result
the parser complete successfuly with the remaining Chunk

ParseMore (Chunk input -> Result input result)
the parser needs more input, pass an empty Chunk or mempty to tell the parser you don't have anymore inputs.

data ParseError input

common parser error definition

Constructors

NotEnough (CountOf (Element input))
meaning the parser was short of CountOf Element of input.

NotEnoughParseOnly
The parser needed more data, only when using parseOnly

ExpectedElement (Element input) (Element input)
when using element

Expected (Chunk input) (Chunk input)
when using elements or string

Satisfy (Maybe String)
the satisfy or satisfy_ function failed,

reportError :: ParseError input -> Parser input a

helper function to report error when writing parsers

This way we can provide more detailed error when building custom parsers and still avoid to use the naughty _fail_.

myParser :: Parser input Int
myParser = reportError $ Satisfy (Just "this function is not implemented...")

Parser source

class (Sequential input, IndexedCollection input) => ParserSource input where

Associated Types

type Chunk input

Methods

nullChunk :: input -> Chunk input -> Bool

appendChunk :: input -> Chunk input -> input

subChunk :: input -> Offset (Element input) -> CountOf (Element input) -> Chunk input

spanChunk :: input -> Offset (Element input) -> (Element input -> Bool) -> (Chunk input, Offset (Element input))

combinator

peek :: ParserSource input => Parser input (Maybe (Element input))

peek the first element from the input source without consuming it

Returns Nothing if there is no more input to parse.

element :: (ParserSource input, Eq (Element input), Element input ~ Element (Chunk input)) => Element input -> Parser input ()

anyElement :: ParserSource input => Parser input (Element input)

Get the next `Element input` from the parser

elements :: (ParserSource input, Sequential (Chunk input), Element (Chunk input) ~ Element input, Eq (Chunk input)) => Chunk input -> Parser input ()

string :: String -> Parser String ()

satisfy :: ParserSource input => Maybe String -> (Element input -> Bool) -> Parser input (Element input)

take one element if satisfy the given predicate

satisfy_ :: ParserSource input => (Element input -> Bool) -> Parser input (Element input)

take one element if satisfy the given predicate

take :: (ParserSource input, Sequential (Chunk input), Element input ~ Element (Chunk input)) => CountOf (Element (Chunk input)) -> Parser input (Chunk input)

takeWhile :: (ParserSource input, Sequential (Chunk input)) => (Element input -> Bool) -> Parser input (Chunk input)

takeAll :: (ParserSource input, Sequential (Chunk input)) => Parser input (Chunk input)

Take the remaining elements from the current position in the stream

skip :: ParserSource input => CountOf (Element input) -> Parser input ()

skipWhile :: (ParserSource input, Sequential (Chunk input)) => (Element input -> Bool) -> Parser input ()

skipAll :: (ParserSource input, Collection (Chunk input)) => Parser input ()

consume every chunk of the stream

(<|>) :: Alternative f => f a -> f a -> f a infixl 3

An associative binary operation

many :: Alternative f => f a -> f [a]

Zero or more.

some :: Alternative f => f a -> f [a]

One or more.

optional :: Alternative f => f a -> f (Maybe a)

One or none.

It is useful for modelling any computation that is allowed to fail.

Examples

Expand

Using the Alternative instance of Control.Monad.Except, the following functions:

>>> import Control.Monad.Except

>>> canFail = throwError "it failed" :: Except String Int
>>> final = return 42                :: Except String Int

Can be combined by allowing the first function to fail:

>>> runExcept $ canFail *> final
Left "it failed"
>>> runExcept $ optional canFail *> final
Right 42

repeat :: ParserSource input => Condition -> Parser input a -> Parser input [a]

repeat the given parser a given amount of time

Unlike some or many, this operation will bring more precision on how many times you wish a parser to be sequenced.

## Repeat Exactly a number of time

repeat (Exactly 6) (takeWhile ((/=) ',') <* element ',')

## Repeat Between lower `And` upper times

repeat (Between $ 1 `And` 10) (takeWhile ((/=) ',') <* element ',')

data Condition

Constructors

Between !And

Exactly !Word

data And

Constructors

And !Word !Word