Foundation.Primitive

Description

Documentation

class Eq ty => PrimType ty where

Represent the accessor for types that can be stored in the UArray and MUArray.

Types need to be a instance of storable and have fixed sized.

Associated Types

type PrimSize ty :: Nat

type level size of the given ty

Methods

primSizeInBytes :: Proxy ty -> CountOf Word8

get the size in bytes of a ty element

primShiftToBytes :: Proxy ty -> Int

get the shift size

primBaUIndex :: ByteArray# -> Offset ty -> ty

return the element stored at a specific index

primMbaURead

Arguments

:: PrimMonad prim

=> MutableByteArray# (PrimState prim)
mutable array to read from

-> Offset ty
index of the element to retrieve

-> prim ty
the element returned

Read an element at an index in a mutable array

primMbaUWrite

Arguments

:: PrimMonad prim

=> MutableByteArray# (PrimState prim)
mutable array to modify

-> Offset ty
index of the element to modify

-> ty
the new value to store

-> prim ()

Write an element to a specific cell in a mutable array.

primAddrIndex :: Addr# -> Offset ty -> ty

Read from Address, without a state. the value read should be considered a constant for all pratical purpose, otherwise bad thing will happens.

primAddrRead :: PrimMonad prim => Addr# -> Offset ty -> prim ty

Read a value from Addr in a specific primitive monad

primAddrWrite :: PrimMonad prim => Addr# -> Offset ty -> ty -> prim ()

Write a value to Addr in a specific primitive monad

class (Functor m, Applicative m, Monad m) => PrimMonad (m :: Type -> Type) where

Primitive monad that can handle mutation.

For example: IO and ST.

Associated Types

type PrimState (m :: Type -> Type)

type of state token associated with the PrimMonad m

type PrimVar (m :: Type -> Type) :: Type -> Type

type of variable associated with the PrimMonad m

Methods

primitive :: (State# (PrimState m) -> (# State# (PrimState m), a #)) -> m a

Unwrap the State# token to pass to a function a primitive function that returns an unboxed state and a value.

primThrow :: Exception e => e -> m a

Throw Exception in the primitive monad

unPrimMonad :: m a -> State# (PrimState m) -> (# State# (PrimState m), a #)

Run a Prim monad from a dedicated state#

primVarNew :: a -> m (PrimVar m a)

Build a new variable in the Prim Monad

primVarRead :: PrimVar m a -> m a

Read the variable in the Prim Monad

primVarWrite :: PrimVar m a -> a -> m ()

Write the variable in the Prim Monad

endianess

class ByteSwap a

Class of types that can be byte-swapped.

e.g. Word16, Word32, Word64

Minimal complete definition

byteSwap

newtype LE a

Little Endian value

Constructors

LE

Fields

unLE :: a

toLE :: ByteSwap a => a -> LE a

Convert a value in cpu endianess to little endian

fromLE :: ByteSwap a => LE a -> a

Convert from a little endian value to the cpu endianness

newtype BE a

Big Endian value

Constructors

BE

Fields

unBE :: a

toBE :: ByteSwap a => a -> BE a

Convert a value in cpu endianess to big endian

fromBE :: ByteSwap a => BE a -> a

Convert from a big endian value to the cpu endianness

Integral convertion

class IntegralUpsize a b where

Upsize an integral value

The destination type b size need to be greater or equal than the size type of a

Methods

integralUpsize :: a -> b

class IntegralDownsize a b where

Downsize an integral value

Minimal complete definition

integralDownsizeCheck

Methods

integralDownsize :: a -> b

integralDownsizeCheck :: a -> Maybe b

Evaluation

class NormalForm a where

Data that can be fully evaluated in Normal Form

Methods

toNormalForm :: a -> ()

force :: NormalForm a => a -> a

deepseq :: NormalForm a => a -> b -> b

These

data These a b

Either a or b or both.

Constructors

This a

That b

These a b

Block of memory

data Block ty

A block of memory containing unpacked bytes representing values of type ty

data MutableBlock ty st

A Mutable block of memory containing unpacked bytes representing values of type ty

Ascii

data Char7

ASCII value between 0x0 and 0x7f

data AsciiString

Opaque packed array of characters in the ASCII encoding