Foundation.String.ASCII

Description

A AsciiString type backed by a ASCII encoded byte array and all the necessary functions to manipulate the string.

The recommended type is AsciiString from UTF8

Documentation

data AsciiString

Opaque packed array of characters in the ASCII encoding

create :: PrimMonad prim => CountOf CUChar -> (MutableAsciiString (PrimState prim) -> prim (Offset CUChar)) -> prim AsciiString

replicate :: CountOf CUChar -> CUChar -> AsciiString

Binary conversion

fromBytesUnsafe :: UArray CUChar -> AsciiString

Convert a Byte Array representing UTF8 data directly to a string without checking for UTF8 validity

If the input contains invalid sequences, it will trigger runtime async errors when processing data.

In doubt, use fromBytes

toBytes :: AsciiString -> UArray CUChar

copy :: AsciiString -> AsciiString

Copy the AsciiString

Legacy utility

lines :: AsciiString -> [AsciiString]

words :: AsciiString -> [AsciiString]