Foundation.System.Info

Description

Operation System info

data OS

Constructors

Windows

OSX

Linux

Android

BSD

os :: Either [Char] OS

get the operating system on which the program is running.

Either return the known OS or a strict String of the OS name.

This function uses the base's os function.

CPU info

data Arch

Enumeration of the known GHC supported architecture.

Constructors

I386

X86_64

PowerPC

PowerPC64

Sparc

Sparc64

ARM

ARM64

arch :: Either [Char] Arch

get the machine architecture on which the program is running

Either return the known architecture or a Strict String of the architecture name.

This function uses the base's arch function.

cpus :: IO Int

returns the number of CPUs the machine has

data Endianness

Constructors

LittleEndian

BigEndian

endianness :: Endianness

endianness of the current architecture

Compiler info

compilerName :: String

get the compiler name

get the compilerName from base package but convert it into a strict String

compilerVersion :: Version

The version of compilerName with which the program was compiled or is being interpreted.

Example

Expand

ghci> compilerVersion
Version {versionBranch = [8,8], versionTags = []}

data Version

A Version represents the version of a software entity.

An instance of Eq is provided, which implements exact equality modulo reordering of the tags in the versionTags field.

An instance of Ord is also provided, which gives lexicographic ordering on the versionBranch fields (i.e. 2.1 > 2.0, 1.2.3 > 1.2.2, etc.). This is expected to be sufficient for many uses, but note that you may need to use a more specific ordering for your versioning scheme. For example, some versioning schemes may include pre-releases which have tags "pre1", "pre2", and so on, and these would need to be taken into account when determining ordering. In some cases, date ordering may be more appropriate, so the application would have to look for date tags in the versionTags field and compare those. The bottom line is, don't always assume that compare and other Ord operations are the right thing for every Version.

Similarly, concrete representations of versions may differ. One possible concrete representation is provided (see showVersion and parseVersion), but depending on the application a different concrete representation may be more appropriate.

Constructors

Version

Fields

versionBranch :: [Int]

The numeric branch for this version. This reflects the fact that most software versions are tree-structured; there is a main trunk which is tagged with versions at various points (1,2,3...), and the first branch off the trunk after version 3 is 3.1, the second branch off the trunk after version 3 is 3.2, and so on. The tree can be branched arbitrarily, just by adding more digits.

We represent the branch as a list of Int, so version 3.2.1 becomes [3,2,1]. Lexicographic ordering (i.e. the default instance of Ord for [Int]) gives the natural ordering of branches.

versionTags :: [String]

A version can be tagged with an arbitrary list of strings. The interpretation of the list of tags is entirely dependent on the entity that this version applies to.