Foundation.Time.StopWatch

Documentation

data StopWatchPrecise

A precise stop watch

The precision is higher than a normal stopwatch, but also on some system it might not be able to record longer period of time accurately (possibly wrapping)

startPrecise :: IO StopWatchPrecise

Create a new precise stop watch

record the time at start of call

stopPrecise :: StopWatchPrecise -> IO NanoSeconds

Get the number of nano seconds since the call to startPrecise