Foundation.Array

Description

Simple Array and Almost-Array-like data structure

Generally accessible in o(1)

Documentation

data Array a :: * -> *

Array of a

data MArray a st :: * -> * -> *

Mutable Array of a

data UArray ty :: * -> *

An array of type built on top of GHC primitive.

The elements need to have fixed sized and the representation is a packed contiguous array in memory that can easily be passed to foreign interface

data MUArray ty st :: * -> * -> *

A Mutable array of types built on top of GHC primitive.

Element in this array can be modified in place.

data ChunkedUArray ty

data Bitmap

data MutableBitmap st

class Eq ty => PrimType ty

Represent the accessor for types that can be stored in the UArray and MUArray.

Types need to be a instance of storable and have fixed sized.

Minimal complete definition

primSizeInBytes, primShiftToBytes, primBaUIndex, primMbaURead, primMbaUWrite, primAddrIndex, primAddrRead, primAddrWrite

data OutOfBound :: *

Exception during an operation accessing the vector out of bound

Represent the type of operation, the index accessed, and the total length of the vector.