Foundation.Array.Internal

Description

Give access to Array non public functions which can be used to make certains optimisations.

Most of what is available here has no guarantees of stability. Anything can be removed and changed.

Documentation

data UArray ty

An array of type built on top of GHC primitive.

The elements need to have fixed sized and the representation is a packed contiguous array in memory that can easily be passed to foreign interface

Constructors

UVecBA !(Offset ty) !(Size ty) !PinnedStatus ByteArray#

UVecAddr !(Offset ty) !(Size ty) !(FinalPtr ty)

fromForeignPtr :: PrimType ty => (ForeignPtr ty, Int, Int) -> UArray ty

withPtr :: (PrimMonad prim, PrimType ty) => UArray ty -> (Ptr ty -> prim a) -> prim a

recast :: (PrimType a, PrimType b) => UArray a -> UArray b

toHexadecimal :: PrimType ty => UArray ty -> UArray Word8