Foundation.Conduit.Textual

Documentation

lines :: Monad m => Conduit String String m ()

Split conduit of string to its lines

This is very similar to Prelude lines except it work directly on Conduit

Note that if the newline character is not ever appearing in the stream, this function will keep accumulating data until OOM

TODO: make a size-limited function

words :: Monad m => Conduit String String m ()

fromBytes :: MonadThrow m => Encoding -> Conduit (UArray Word8) String m ()

toBytes :: Monad m => Encoding -> Conduit String (UArray Word8) m ()