Foundation.Format.CSV

Description

Provies the support for Comma Separated Value

CSV

data CSV

CSV Type

Builder

String Bulider

csvStringBuilder :: CSV -> Builder

serialise the CSV document into a UTF8 string

rowStringBuilder :: Row -> Builder

fieldStringBuilder :: Field -> Builder

Block Builder

csvBlockBuilder :: CSV -> Builder

serialise the CSV document into a UTF8 encoded (Block Word8)

rowBlockBuilder :: Row -> Builder

fieldBlockBuilder :: Field -> Builder

Conduit

rowC :: (ToRow row, Monad m) => Conduit row (Block Word8) m ()

Row

data Row

CSV Row

class ToRow a where

Minimal complete definition

toRow

Methods

toRow :: a -> Row

Field

data Field

CSV field

Constructors

FieldInteger Integer

FieldDouble Double

FieldString String Escaping

data Escaping

Constructors

NoEscape

Escape

DoubleEscape

class ToField a where

Minimal complete definition

toField

Methods

toField :: a -> Field

helpers

integral :: Into Integer a => a -> Field

helper function to create a FieldInteger

float :: Double -> Field

string :: String -> Field

heler function to create a FieldString.

This function will findout automatically if an escaping is needed. if you wish to perform the escaping manually, do not used this function