Foundation.IO

Description

IO Routine

Terminal

putStrLn :: String -> IO ()

Print a string with a newline to standard output

putStr :: String -> IO ()

Print a string to standard output

stdin :: Handle

A handle managing input from the Haskell program's standard input channel.

stdout :: Handle

A handle managing output to the Haskell program's standard output channel.

File

data IOMode :: *

See openFile

Constructors

ReadMode

WriteMode

AppendMode

ReadWriteMode

openFile :: FilePath -> IOMode -> IO Handle

list the file name in the given FilePath directory

TODO: error management and not implemented yet getDirectory :: FilePath -> IO [FileName] getDirectory = undefined

Open a new handle on the file

closeFile :: Handle -> IO ()

Close a handle

withFile :: FilePath -> IOMode -> (Handle -> IO r) -> IO r

withFile filepath mode act opens a file using the mode and run act. the by-product handle will be closed when act finish, either normally or through an exception.

The value returned is the result of act@

hGet :: Handle -> Int -> IO (UArray Word8)

Read binary data directly from the specified Handle.

First argument is the Handle to read from, and the second is the number of bytes to read. It returns the bytes read, up to the specified size, or an empty array if EOF has been reached.

hGet is implemented in terms of hGetBuf.

hPut :: Handle -> UArray Word8 -> IO ()

readFile :: FilePath -> IO (UArray Word8)

Read a binary file and return the whole content in one contiguous buffer.

foldTextFile

Arguments

:: (String -> a -> IO a)
Fold callback function

-> a
initial accumulator

-> FilePath
File to read

-> IO a

Deprecated: use conduit instead

Fold over chunks file calling the callback function for each chunks read from the file, until the end of file.