Foundation.Monad

Documentation

class Monad m => MonadIO (m :: * -> *) where

Monads in which IO computations may be embedded. Any monad built by applying a sequence of monad transformers to the IO monad will be an instance of this class.

Instances should satisfy the following laws, which state that liftIO is a transformer of monads:

liftIO . return = return

liftIO (m >>= f) = liftIO m >>= (liftIO . f)

Minimal complete definition

liftIO

Methods

liftIO :: IO a -> m a

Lift a computation from the IO monad.

class AMPMonad m => MonadFailure (m :: * -> *) where

Monad that can represent failure

Similar to MonadFail but with a parametrized Failure linked to the Monad

Minimal complete definition

mFail

Associated Types

type Failure (m :: * -> *) :: *

The associated type with the MonadFailure, representing what failure can be encoded in this monad

Methods

mFail :: Failure m -> m ()

Raise a Failure through a monad.

class AMPMonad m => MonadThrow m where

Monad that can throw exception

Minimal complete definition

throw

Methods

throw :: Exception e => e -> m a

Throw immediatity an exception. Only a MonadCatch monad will be able to catch the exception using catch

class MonadThrow m => MonadCatch m where

Monad that can catch exception

Minimal complete definition

catch

Methods

catch :: Exception e => m a -> (e -> m a) -> m a

class MonadCatch m => MonadBracket m where

Monad that can ensure cleanup actions are performed even in the case of exceptions, both synchronous and asynchronous. This usually excludes continuation-based monads.

Minimal complete definition

generalBracket

Methods

generalBracket

Arguments

:: m a
acquire some resource

-> (a -> b -> m ignored1)
cleanup, no exception thrown

-> (a -> SomeException -> m ignored2)
cleanup, some exception thrown. The exception will be rethrown

-> (a -> m b)
inner action to perform with the resource

-> m b

A generalized version of the standard bracket function which allows distinguishing different exit cases.

class MonadTrans trans where

Basic Transformer class

Minimal complete definition

lift

Methods

lift :: AMPMonad m => m a -> trans m a

Lift a computation from an inner monad to the current transformer monad

newtype Identity a :: * -> *

Identity functor and monad. (a non-strict monad)

Since: 4.8.0.0

Constructors

Identity

Fields

runIdentity :: a

replicateM :: Applicative m => CountOf a -> m a -> m [a]

replicateM n act performs the action n times, gathering the results.