Foundation.Parser

Description

The current implementation is mainly, if not copy/pasted, inspired from memory's Parser.

A very simple bytearray parser related to Parsec and Attoparsec

Simple example:

> parse ((,,) <$> take 2 <*> element 0x20 <*> (elements "abc" *> anyElement)) "xx abctest"
ParseOK "est" ("xx", 116)

Documentation

newtype Parser input a

Simple parser structure

Constructors

Parser

Fields

runParser :: forall r. input -> Failure input r -> Success input a r -> Result input r

data Result input a

Simple parsing result, that represent respectively:

failure: with the error message

continuation: that need for more input data

success: the remaining unparsed data and the parser value

Constructors

ParseFail (ParserError input)

ParseMore (Maybe input -> Result input a)

ParseOK input a

data ParserError input

Constructors

Expected

Fields

expectedInput :: !input

the expected input

receivedInput :: !input

but received this data

DoesNotSatify
some bytes didn't satisfy predicate

NotEnough
not enough data to complete the parser

MonadFail String
only use in the event of Monad.fail function

run the Parser

parse :: Sequential input => Parser input a -> input -> Result input a

Run a Parser on a ByteString and return a Result

parseFeed :: (Sequential input, Monad m) => m (Maybe input) -> Parser input a -> input -> m (Result input a)

Run a parser on an @initial input.

If the Parser need more data than available, the @feeder function is automatically called and fed to the More continuation.

parseOnly :: (Typeable input, Show input, Sequential input, Element input ~ Char) => Parser input a -> input -> a

parse only the given input

The left-over `Element input` will be ignored, if the parser call for more data it will be continuously fed with Nothing (up to 256 iterations).

Parser methods

hasMore :: Sequential input => Parser input Bool

element :: (Sequential input, Eq (Element input)) => Element input -> Parser input ()

Parse a specific `Element input` at current position

if the `Element input` is different than the expected one, this parser will raise a failure.

satisfy :: Sequential input => (Element input -> Bool) -> Parser input (Element input)

take one element if satisfy the given predicate

anyElement :: Sequential input => Parser input (Element input)

Get the next `Element input` from the parser

elements :: (Show input, Eq input, Sequential input) => input -> Parser input ()

Parse a sequence of elements from current position

if the following `Element input` don't match the expected input completely, the parser will raise a failure

string :: String -> Parser String ()

take :: Sequential input => Int -> Parser input input

Take @n elements from the current position in the stream

takeWhile :: Sequential input => (Element input -> Bool) -> Parser input input

Take elements while the @predicate hold from the current position in the stream

takeAll :: Sequential input => Parser input input

Take the remaining elements from the current position in the stream

skip :: Sequential input => Int -> Parser input ()

Skip @n elements from the current position in the stream

skipWhile :: Sequential input => (Element input -> Bool) -> Parser input ()

Skip `Element input` while the @predicate hold from the current position in the stream

skipAll :: Sequential input => Parser input ()

Skip all the remaining `Element input` from the current position in the stream

utils

optional :: Alternative f => f a -> f (Maybe a)

One or none.

many :: Alternative f => forall a. f a -> f [a]

Zero or more.

some :: Alternative f => forall a. f a -> f [a]

One or more.

(<|>) :: Alternative f => forall a. f a -> f a -> f a

An associative binary operation

data Count

Constructors

Never

Once

Twice

Other Int

data Condition

Constructors

Exactly Count

Between Count Count

repeat :: Sequential input => Condition -> Parser input a -> Parser input [a]

repeat the given Parser a given amount of time

If you know you want it to exactly perform a given amount of time:

``` repeat (Exactly Twice) (element a) ```

If you know your parser must performs from 0 to 8 times:

``` repeat (Between Never (Other 8)) ```

This interface is still WIP* but went handy when writting the IPv4/IPv6 parsers.