Foundation.String.Builder

Description

String Builder

This is extremely bad implementation of a builder implementation but provide a very similar API to the future fast implementation; So, in the spirit of getting started and to be able to start using the API, we don't wait for the fast implementation.

Documentation

data Builder

emit :: String -> Builder

emitChar :: Char -> Builder

toString :: Builder -> String